requirements.txt
openai>=1.76.0
python-dotenv
flask
flask-cors
